Separate api handlers and routes
