+++ lib/routes/api/accounts/index.js
@@ -5,7 +5,6 @@ const { User } = require('lib/models')
 const schema = require('./schema')
 const router = express.Router()
 
-// create account [External/Internal Account|mongoDB]
 router.post('/register', async (req, res, next) => {
   const { error, value } = schema.register(req.body)
   var register = value
