+++ lib/routes/api/banks/handler.js
@@ -5,7 +5,7 @@ const openPayments = require('lib/external/open_payments')
 const { Org, Bank } = require('lib/models')
 const schema = require('./schema')
 
-exports.show = async function show (req, res, next) {
+exports.find = async function find (req, res, next) {
   const organization = await Org.findOne({ users: req.user.id }).exec()
   if (organization) {
     let bank = await Bank.findOne({ organization: organization.id }).exec()
