+++ lib/routes/api/banks/index.js
@@ -6,7 +6,7 @@ module.exports = [
     method: 'GET',
     path: '/',
     authorizors: 'loggedIn',
-    handler: handler.show
+    handler: handler.find
   },
   {
     description: 'Create bank',
