+++ lib/routes/api/users/handler.js
@@ -1,0 +1,3 @@
+exports.findAll = async function findAll (req, res, next) {
+  return res.send([])
+}
