+++ lib/routes/api/users/index.js
@@ -1,10 +1,11 @@
+const handler = require('./handler')
 
-const express = require('express')
-
-const router = express.Router()
-
-router.get('/list', async (req, res, next) => {
-  return res.send([])
-})
-
-module.exports = router
+module.exports = [
+  {
+    description: 'Get all users',
+    method: 'GET',
+    path: '/list',
+    authorizors: 'loggedIn',
+    handler: handler.findAll
+  }
+]
