+++ server.js
@@ -54,6 +54,13 @@ server.use(lusca({
 server.use(routes)
 
 server.use('/api/banks', router(require('lib/routes/api/banks')))
+server.use('/api/cards', router(require('lib/routes/api/cards')))
+server.use('/api/organizations', router(require('lib/routes/api/organizations')))
+server.use('/api/users', router(require('lib/routes/api/users')))
+server.use('/api/contacts', router(require('lib/routes/api/contacts')))
+server.use('/api/history', router(require('lib/routes/api/history')))
+server.use('/api/newsletter', router(require('lib/routes/api/newsletter')))
+server.use('/api/projects', router(require('lib/routes/api/projects')))
 
 server.get('/*', function (req, res) {
   res.sendFile(path.join(__dirname, './build', 'index.html'))
